requirements.txt
torch
torchvision
timm
streamlit
Pillow
numpy
pyyaml
scipy